Add FrameworkReference and fix ViewCount type
- Add Microsoft.AspNetCore.App framework reference for Identity, HttpClient
- Add Microsoft.Extensions.Http and Options.ConfigurationExtensions
- Change ViewCount from int to long in NewsSearchDocument

+++ apps/api/CommonHall.Infrastructure/Search/SearchDocuments.cs
@@ -18,7 +18,7 @@ public record NewsSearchDocument
     public DateTimeOffset CreatedAt { get; init; }
     public DateTimeOffset UpdatedAt { get; init; }
     public required string Slug { get; init; }
-    public int ViewCount { get; init; }
+    public long ViewCount { get; init; }
     public object? VisibilityRule { get; init; }
     public string? TeaserImageUrl { get; init; }
 }
